moving test server into separate file
